pyproject.toml
[project]
name = "sessao-certa-api"
version = "0.1.0"
description = ""
authors = [
    {name = "RICARDO",email = "ricardomarquesddsa@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standart] (>=0.123.9,<0.124.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "pydantic-extra-types (>=2.11.0,<3.0.0)",
    "phonenumbers (>=9.0.21,<10.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [".", "src"]
